requirements-test.txt
# ================================
# Testing Dependencies
# ================================

# Core Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# Mocking and Fixtures
pytest-mock>=3.11.0
faker>=19.0.0

# HTTP Testing
requests>=2.31.0
responses>=0.23.0

# Coverage Reporting
coverage>=7.3.0

# Load Testing
locust>=2.15.0
